+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Improvements
 - Allow changing the entire ui style for a single element
 - Skip unnecessary area updates for elements with dirty parents
 - Calculate panel scroll bar height based on content height
+- Remember the location that a scroll bar scroller was grabbed in when scrolling
 
 Fixes
 - Fixed VerticalSpace height parameter being an integer